+++ packages/dgt-components/lib/components/authentication/authenticate.component.ts
@@ -3,7 +3,6 @@ import { createMachine, DoneEvent, interpret, Interpreter, State, StateMachine }
 import { RxLitElement } from 'rx-lit';
 import { from } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { getLoggerFor, Logger } from '@digita-ai/handlersjs-logging';
 import { Theme, DigitaBlue } from '@digita-ai/dgt-theme';
 import { SolidService } from '@digita-ai/inrupt-solid-service';
 import { Translator } from '@digita-ai/dgt-utils';
@@ -19,13 +18,15 @@ export class AuthenticateComponent extends RxLitElement {
 
   private actor: Interpreter<AuthenticateContext, AuthenticateStateSchema, AuthenticateEvent, AuthenticateState>;
   private machine: StateMachine<AuthenticateContext, AuthenticateStateSchema, AuthenticateEvent, AuthenticateState>;
-  private logger: Logger =  getLoggerFor(this, 5, 5);
 
   @internalProperty()
   state?: State<AuthenticateContext>;
 
   @internalProperty()
   issuers?: Issuer[];
+
+  @internalProperty()
+  validating = false;
 
   @property({ type: Boolean }) hideWebId = false;
   @property({ type: Boolean }) hideIssuers = false;
@@ -67,6 +68,7 @@ export class AuthenticateComponent extends RxLitElement {
 
     this.subscribe('state', from(this.actor));
     this.subscribe('issuers', from(this.actor).pipe(map((state) => state.context.issuers)));
+    this.subscribe('validating', from(this.actor).pipe(map((state) => state.hasTag('validating'))));
 
     this.subscribe('webIdValidationResults', from(this.actor).pipe(map((state) => {
 
@@ -101,7 +103,6 @@ export class AuthenticateComponent extends RxLitElement {
 
   onSubmit = (event: CustomEvent): void => {
 
-    this.logger.info('onSubmit', event);
     event.preventDefault();
     this.actor.send(new ClickedLoginEvent(event.detail));
 
@@ -109,7 +110,6 @@ export class AuthenticateComponent extends RxLitElement {
 
   onWebIdChange = (event: CustomEvent): void => {
 
-    this.logger.info('onWebIdChange', event);
     event.preventDefault();
     this.actor.send(new WebIdEnteredEvent(event.detail));
 
@@ -117,7 +117,6 @@ export class AuthenticateComponent extends RxLitElement {
 
   onButtonCreateWebIDClick = (): void => {
 
-    this.logger.info('onButtonCreateWebIDClick');
     this.dispatchEvent(new CustomEvent('create-webid', { bubbles: true }));
 
   };
@@ -157,13 +156,17 @@ export class AuthenticateComponent extends RxLitElement {
           .textButton="${this.textButton}"
           .validationResults="${this.webIdValidationResults}"
           .translator="${this.translator}"
+          ?validating="${this.validating}"
           .layout="${this.layout}"
         >
           <slot name="beforeWebId" slot="before"></slot>
           <slot name="afterWebId" slot="after"></slot>
         </webid-form>
        ` : html` ${ this.state?.matches(AuthenticateStates.SELECTING_ISSUER) ? html`
-        <provider-list @issuer-selected="${(event: CustomEvent) => this.actor.send(new SelectedIssuerEvent(event.detail))}" .providers="${this.issuers}"></provider-list>`
+        <provider-list @issuer-selected="${(event: CustomEvent) => this.actor.send(new SelectedIssuerEvent(event.detail))}" .providers="${this.issuers}">
+          <slot name="beforeIssuers" slot="before"></slot>
+          <slot name="afterIssuers" slot="after"></slot>
+        </provider-list>`
     : html`<loading-component part="loading"></loading-component>` }
     `}`;