+++ packages/dgt-components/lib/components/authentication/webid.component.ts
@@ -5,9 +5,9 @@ import { Theme } from '@digita-ai/dgt-theme';
 import { debounce } from 'debounce';
 import { unsafeSVG } from 'lit-html/directives/unsafe-svg';
 import { Translator } from '@digita-ai/dgt-utils';
-import { getLoggerFor, Logger } from '@digita-ai/handlersjs-logging';
 import { define } from '../../util/define';
 import { AlertComponent } from '../alerts/alert.component';
+import { LoadingComponent } from '../loading/loading.component';
 
 export class WebIdComponent extends RxLitElement {
 
@@ -19,23 +19,21 @@ export class WebIdComponent extends RxLitElement {
   @property({ type: Array }) validationResults: string[];
   @property({ type: Boolean }) hideCreateNewWebId = false;
   @property({ type: Boolean }) disableLogin = true; // disable button by default
+  @property({ type: Boolean }) validating = false;
   @property({ type: Translator }) translator?: Translator;
-
-  private logger: Logger = getLoggerFor(this, 5, 5);
 
   constructor() {
 
     super();
 
     define('alert-component', AlertComponent);
+    define('loading-component', LoadingComponent);
 
   }
 
   onSubmit = (event: Event & { target: HTMLFormElement }): void => {
 
     event.preventDefault();
-
-    this.logger.info('onSubmit', event.target);
 
     this.dispatchEvent(new CustomEvent('submit-webid', {
       detail: event.target.querySelector<HTMLInputElement>('input[name=webid]').value,
@@ -45,8 +43,6 @@ export class WebIdComponent extends RxLitElement {
 
   onWebIdChange = debounce((target: HTMLInputElement): void => {
 
-    this.logger.info('onWebIdChange', target);
-
     this.dispatchEvent(new CustomEvent('change-webid', {
       detail: target.value,
     }));
@@ -55,7 +51,6 @@ export class WebIdComponent extends RxLitElement {
 
   onButtonCreateWebIDClick = (): void => {
 
-    this.logger.info('onButtonCreateWebIDClick');
     this.dispatchEvent(new CustomEvent('create-webid'));
 
   };
@@ -82,7 +77,10 @@ export class WebIdComponent extends RxLitElement {
 
         <div class="webid-input-button-container ${ this.layout }" part="webid-input-button-container">
           <div class="webid-input-container" part="webid-input-container">
-            <input part="webid-input" type="text" id="webid" name="webid" placeholder="${this.textPlaceholder}" @input="${(event: InputEvent) => { this.onWebIdChange(event.target as HTMLInputElement); }}"/>
+            <div class="input-container">
+              <input part="webid-input" type="text" id="webid" name="webid" placeholder="${this.textPlaceholder}" @input="${(event: InputEvent) => { this.onWebIdChange(event.target as HTMLInputElement); }}"/>
+              <loading-component ?hidden="${!this.validating}" part="loading"></loading-component>
+            </div>
             ${ this.layout === 'vertical' ? alerts : ''}
           </div>  
           <button
@@ -116,7 +114,6 @@ export class WebIdComponent extends RxLitElement {
           flex-direction: column;
           gap: var(--gap-normal);
         }
-
         button {
           width: 100%;
           border-radius: var(--border-large);
@@ -148,9 +145,20 @@ export class WebIdComponent extends RxLitElement {
           align-items: center;
           justify-content: center;
         }
+        .webid-input-button-container button:hover:enabled {
+          background-color: var(--colors-primary-light);
+          border: 2px solid var(--colors-primary-light);
+        }
         input  {
           padding: var(--gap-normal);
+          width: 100%;
+          border: none;
         }
+
+        input:focus-visible{
+          outline: none;
+        }
+
         a {
           font-size: var(--font-size-small);
           padding: var(--gap-tiny);
@@ -159,11 +167,11 @@ export class WebIdComponent extends RxLitElement {
           align-self: flex-end;
           cursor: pointer;
         }
-
+        
         a:hover {
           color: var(--colors-primary-dark);
         }
-
+        
         h1 {
           margin: var(--gap-large) var(--gap-normal);
           font-size: var(--font-size-header-normal);
@@ -171,7 +179,21 @@ export class WebIdComponent extends RxLitElement {
           font-weight: bold;
           text-align: center;
         }
+        
+        .input-container {
+          display: flex;
+          align-items: center;
+          background-color: var(--colors-background-light);
+        }
 
+        loading-component {
+          height: var(--gap-normal);
+          width: var(--gap-normal);
+          display: block;
+          padding-right: var(--gap-normal);
+          flex: 1 0;
+        }
+        
         `,
     ];