+++ packages/dgt-components/lib/components/forms/form-element.component.ts
@@ -6,7 +6,6 @@ import { RxLitElement } from 'rx-lit';
 import { from } from 'rxjs';
 import { map, tap } from 'rxjs/operators';
 import { Loading, Theme } from '@digita-ai/dgt-theme';
-import { getLoggerFor } from '@digita-ai/handlersjs-logging';
 import { FormContext, FormRootStates, FormState, FormStateSchema, FormSubmissionStates, FormValidationStates } from './form.machine';
 import { FormValidatorResult } from './form-validator-result';
 import { FormEvent, FormEvents, FormUpdatedEvent } from './form.events';
@@ -16,13 +15,11 @@ import { FormEvent, FormEvents, FormUpdatedEvent } from './form.events';
  */
 export class FormElementComponent<T> extends RxLitElement {
 
-  private logger = getLoggerFor(this, 5, 5);
-
   /**
    * All input elements slotted in the form element.
    */
   @internalProperty()
-  inputs: HTMLInputElement[];
+  inputs: (HTMLInputElement | HTMLTextAreaElement)[];
 
   /**
    * The slot element which contains the input field.
@@ -141,8 +138,6 @@ export class FormElementComponent<T> extends RxLitElement {
 
   protected firstUpdated(changed: PropertyValues): void {
 
-    this.logger.info('firstUpdated', changed);
-
     super.firstUpdated(changed);
 
     this.bindActorToInput(this.inputSlot, this.actor, this.field, this.data);
@@ -161,35 +156,31 @@ export class FormElementComponent<T> extends RxLitElement {
 
     if (!slot) {
 
-      this.logger.warn('bindActorToInput', 'No slot found.');
       throw new ArgumentError('Argument slot should be set.', slot);
 
     }
 
     if (!actor) {
 
-      this.logger.warn('bindActorToInput', 'No actor found.');
       throw new ArgumentError('Argument actor should be set.', actor);
 
     }
 
     if (!field) {
 
-      this.logger.warn('bindActorToInput', 'No field found.');
       throw new ArgumentError('Argument field should be set.', field);
 
     }
 
     if (!data) {
 
-      this.logger.warn('bindActorToInput', 'No data found.');
       throw new ArgumentError('Argument data should be set.', data);
 
     }
 
     this.inputs = slot.assignedNodes({ flatten: true })?.filter(
-      (element) => element instanceof HTMLInputElement,
-    ).map((element) => element as HTMLInputElement);
+      (element) => element instanceof HTMLInputElement || element instanceof HTMLTextAreaElement,
+    ).map((element) => element as HTMLInputElement | HTMLTextAreaElement);
 
     this.inputs?.forEach((element) => {
 
@@ -203,11 +194,9 @@ export class FormElementComponent<T> extends RxLitElement {
       // Listen for Enter presses to submit
       if (this.submitOnEnter) {
 
-        element.addEventListener('keypress', (event) => {
+        element.addEventListener('keypress', (event: KeyboardEvent) => {
 
           if (event.key === 'Enter') {
-
-            this.logger.info('bindActorToInput', 'Enter pressed, submitting form');
 
             actor.send({ type: FormEvents.FORM_SUBMITTED });