+++ packages/dgt-utils/lib/i18n/factories/cached-translator-factory.ts
@@ -1,4 +1,3 @@
-import { getLoggerFor } from '@digita-ai/handlersjs-logging';
 import { MemoryTranslator } from '../translators/memory-translator';
 import { TranslatorFactory } from '../models/translator-factory';
 
@@ -8,13 +7,11 @@ import { TranslatorFactory } from '../models/translator-factory';
 export class CachedTranslatorFactory implements TranslatorFactory {
 
   private translations: { [key: string]: string };
-  private logger = getLoggerFor(this, 5, 5);
 
   async createTranslator(language: string): Promise<MemoryTranslator> {
 
     if (!this.translations) {
 
-      this.logger.info('Fetching translations');
       this.translations = await (await fetch(`${window.location.origin}/${language}.json`)).json();
 
     }