+++ packages/dgt-utils/lib/i18n/translators/memory-translator.ts
@@ -1,13 +1,10 @@
-import { registerTranslateConfig, use, get, Values, ValuesCallback, ITranslateConfig, Strings } from '@appnest/lit-translate';
-import { getLoggerFor } from '@digita-ai/handlersjs-logging';
+import { registerTranslateConfig, use, get, Values, ValuesCallback, ITranslateConfig } from '@appnest/lit-translate';
 import { TranslationsLoadedEvent, Translator } from '../models/translator';
 
 /**
  * An implementation of a Translator which stores translations in-memory.
  */
 export class MemoryTranslator extends Translator {
-
-  private logger = getLoggerFor(this, 5, 5);
 
   /**
    * Translates a key to a specific locale.
@@ -24,12 +21,9 @@ export class MemoryTranslator extends Translator {
 
     if (!key) {
 
-      this.logger.warn('No key given');
       throw new Error('Argument key should be set.');
 
     }
-
-    this.logger.info('Translating', key);
 
     return get(key, values, config);